Add nth propery to reader and interframe processing to cl
'nth' can be used to define the first file (in terms of number) to be read from
a path. 'interframe' is a new mode for the cl filter which takes two frames and
executes a kernel with these frames an argument and returns a resulting image
(for example difference of two images etc.).