Decouple iteration and computation
This is the first in a series of patches that finally add mixed, heterogenous
computation to the UFO framework. For this, we now have kernel-style computation
on a filter. That means the process function is called for each work item, that
is composed of all inputs rather than having the filter iterating over the data
stream. As a bonus, writing filters becomes a tad more easier and needs less
code.