Let the scheduler execute the graph not vice versa
Up to now, the UfoGraph object had a method `run' which was used to start the
computation. Internally this object created a scheduler object and passed the
list of filters to it. However, in reality the graph does not need to know
anything about executing as it is purely a representation of data flow.

Hence, the UfoBaseScheduler now has a `run' method that takes two arguments, a
UfoGraph object and a GError location. Moreover, UfoGraph does not have a
"configuration" property anymore as it does not query path information.