Profile CPU usage
This has a caveat though, especially for the FFT and IFFT filters. Because they
wait explicitly on preceding data they essentially block inside the node thus it
looks like they spent that amount of time. We can fix this later on by
subtracting the SYNC time from them but it is hard to know which filter behaves
like that.