Don't map GPUs more than once
In a multi-rooted graph, the mapper traverses from each root in a depth-first
manner and assigns one of the GPUs. This commit fixes two things: 1) the GPU
index went out of bounds but it wasn't noticed because each node was visited
multiple times. 2) A GPU is only assigned once to a visited node and subsequent
visits are ignored.