Add Bash completion script for ufo-launch
The completion script uses ufo-query to determine available plugins and their
respective property names.

The installation is opt-in for two reasons:

1. bash-completion allows completion only for scripts defined in
/usr/share/bash-completions/* or appended to ~/.bash_completions. The former is
not a viable solution if the user installs into his home directory (see 2.) and
the second is a logistic nightmare.

2. There is no straightforward way to ask CMake if an installation path is
writable at all or at least fail gracefully. Thus if the user is unable to
install outside his home directory, installation will fail badly.