Fix segfault with long-running tasks
This change sanitizes problems with long-running tasks. The sleep in the clean
up function was a crude hack to avoid unreffing objects before execution
stopped. This was not noticed with tasks that were faster than the arbitrary 1.5
seconds. Cleaning up after the scheduler actually finished solves this problem
and avoids the unnecessary wait time.