Fix #148: re-write longest path algorithm
In the previous approach, we build a new sub-graph containing nodes
which fulfilled a certain predicate. However, the ownership of the
contained nodes was not stated correctly which caused a lot of ref/unref
operations in other parts of the code which eventually caused segfaults
and other problems related to the task graph expansion we have seen
before.

This approach is simpler and just scans graph recursively building paths
on the go. The heuristic is not perfect though because we use the first
candidate in a list of outgoing branches which might not yield the best
path. However, situations in which branches diverge are much less seldom
than branches that merge.